+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-Change Log:
+# Change Log #
+* Fixed notification permission check for Android 33 and below
+* [WIP] Fixed back button press on permission check activity no closing app
+
+
+# Previous Changes for Major Version #
+
+*V0.2-RC:*
 * Add basic layout for main activity
 * Create activity to check for and request permissions needed by app to function
 * Add foreground service that displays basic overlay on screen